Update LnControlPanel.ino
Hello,
please check my proposed changes: 
I registered that the newer Adafruit-Librariy does not work correctly with the sketch. The change of the Adress-Definition in Line 61 and 62 is to ensure the compatibility with the recent Adafruit-Libraries.
Then i registered in my build that the inputs of the second MCP have no function. In cause of the Division by 2 in Line 151 the Inputs of the second MCP are not registered. After deleting the term "/2" the inputs of the second MCP work correctly.

Luca, i have to thank you for this project. This is exactly what I needed, THANK YOU!

Greetings from Germay,
Johannes